+++ www/packages/types/src/remark-rehype.ts
@@ -267,11 +267,13 @@ export declare type ComponentLinkFixerOptions = {
   filePath?: string
   basePath?: string
   checkLinksType: ComponentLinkFixerLinkType
+  r2BaseUrl?: string
 }
 
 export declare type LocalLinkOptions = {
   filePath?: string
   basePath?: string
+  r2BaseUrl?: string
 }
 
 export type ExpressionJsVarItem = {