+++ packages/design-system/ui/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @medusajs/ui
 
+## 4.1.12
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @medusajs/icons@2.15.2
+
+## 4.1.11
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @medusajs/icons@2.15.1
+
+## 4.1.10
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @medusajs/icons@2.15.0
+
 ## 4.1.9
 
 ### Patch Changes