+++ packages/plugins/loyalty/CHANGELOG.md
@@ -1,5 +1,41 @@
 # @medusajs/loyalty-plugin
 
+## 2.15.2
+
+### Patch Changes
+
+- Updated dependencies [[`2b21d15640ae459386b0acea4c83804c6f502b9d`](https://github.com/medusajs/medusa/commit/2b21d15640ae459386b0acea4c83804c6f502b9d)]:
+  - @medusajs/framework@2.15.2
+  - @medusajs/cli@2.15.2
+  - @medusajs/test-utils@2.15.2
+  - @medusajs/admin-sdk@2.15.2
+  - @medusajs/icons@2.15.2
+  - @medusajs/ui@4.1.12
+
+## 2.15.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @medusajs/test-utils@2.15.1
+  - @medusajs/admin-sdk@2.15.1
+  - @medusajs/cli@2.15.1
+  - @medusajs/framework@2.15.1
+  - @medusajs/icons@2.15.1
+  - @medusajs/ui@4.1.11
+
+## 2.15.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @medusajs/test-utils@2.15.0
+  - @medusajs/framework@2.15.0
+  - @medusajs/cli@2.15.0
+  - @medusajs/admin-sdk@2.15.0
+  - @medusajs/icons@2.15.0
+  - @medusajs/ui@4.1.10
+
 ## 2.14.2
 
 ### Patch Changes